Prefer 'up-to-date' spelling
When we are providing an English error message for users, we should
avoid 'uptodate'.  According to
https://writingexplained.org/up-to-date-hyphenated, when used as an
adjective (i.e. when used as a synonym for 'current') -- which matches
our usage -- the correct spelling is up-to-date.  (Alternatively, when
used as an adverb, the same page points out it is spelled 'up to date'.)
In either case, when providing an English error message for users, we
shouldn't use 'uptodate'.  Correct the spelling.

Reported-by: Jens Schleusener <Jens.Schleusener@fossies.org>
Signed-off-by: Elijah Newren <newren@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>